Created Avatae component and Back Nav
